cnn file and requirements
So far, the data preprocessing is almost done, just need to reformat the 1d text array into a 2d "image". The Google news word2vec model needs to be downloaded and unzipped in the same folder as cnn.ipynb. The link to the file is in cnn.ipynb under the "Tokenizing Political Statements" section. Github would not let me add that file because it is too big (~362 MB).